Refactor Block into rendered and editing components
- Organize key handlers when in block editing mode
- Separate load/save buttons and file pane shortcuts into separate components
- Move block logic specific to the data structure & algos to own file
