+++ src/features/file-pane/file-pane.scss
@@ -1,22 +1,5 @@
-@import "../../index.scss";
-
 .file-pane {
     min-width: 300px;
     background-color: rgba(0,0,0,0.5);
     padding: 20px;
-}
-
-.load-button-input {
-    height: 0;
-    input {
-        visibility: hidden;
-        width: 0;
-    }
-
-    &:before {
-        @extend .btn;
-        @extend .btn-secondary;
-        content: "Load";
-        display: block;
-    }
 }
