+++ src/features/io/io.scss
@@ -1,0 +1,16 @@
+@import "../../index.scss";
+
+.load-button-input {
+    height: 0;
+    input {
+        visibility: hidden;
+        width: 0;
+    }
+
+    &:before {
+        @extend .btn;
+        @extend .btn-secondary;
+        content: "Load";
+        display: block;
+    }
+}
